Create a package object to hold small type and class definitions
